+++ src/Firebase/Messaging.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Kreait\Firebase;
 
+use GuzzleHttp\Promise;
 use Kreait\Firebase\Exception\InvalidArgumentException;
 use Kreait\Firebase\Exception\Messaging\InvalidArgument;
 use Kreait\Firebase\Exception\Messaging\InvalidMessage;
@@ -11,10 +12,14 @@ use Kreait\Firebase\Exception\Messaging\NotFound;
 use Kreait\Firebase\Messaging\ApiClient;
 use Kreait\Firebase\Messaging\CloudMessage;
 use Kreait\Firebase\Messaging\Message;
+use Kreait\Firebase\Messaging\MessageTarget;
+use Kreait\Firebase\Messaging\MulticastSendReport;
 use Kreait\Firebase\Messaging\RegistrationToken;
+use Kreait\Firebase\Messaging\SendReport;
 use Kreait\Firebase\Messaging\Topic;
 use Kreait\Firebase\Messaging\TopicManagementApiClient;
 use Kreait\Firebase\Util\JSON;
+use Psr\Http\Message\ResponseInterface;
 
 class Messaging
 {
@@ -58,6 +63,58 @@ class Messaging
         $response = $this->messagingApi->sendMessage($message);
 
         return JSON::decode((string) $response->getBody(), true);
+    }
+
+    /**
+     * @param array|Message $message
+     * @param string[]|RegistrationToken[] $deviceTokens
+     *
+     * @return MulticastSendReport
+     */
+    public function sendMulticast($message, array $deviceTokens): MulticastSendReport
+    {
+        if (\is_array($message)) {
+            $message = CloudMessage::fromArray($message);
+        }
+
+        if (!($message instanceof Message)) {
+            throw new InvalidArgumentException(
+                'Unsupported message type. Use an array or a class implementing %s'.Message::class
+            );
+        }
+
+        if (!($message instanceof CloudMessage)) {
+            $message = CloudMessage::fromArray($message->jsonSerialize());
+        }
+
+        $promises = [];
+
+        foreach ($deviceTokens as $token) {
+            $target = MessageTarget::with(MessageTarget::TOKEN, (string) $token);
+            $message = $message->withChangedTarget($target->type(), $target->value());
+            $promises[$target->value()] = $this->messagingApi->sendMessageAsync($message);
+        }
+
+        return Promise\settle($promises)
+            ->then(static function (array $results) {
+                $reports = [];
+
+                foreach ($results as $tokenString => $result) {
+                    $target = MessageTarget::with(MessageTarget::TOKEN, $tokenString);
+                    if ($result['state'] === Promise\PromiseInterface::FULFILLED) {
+                        /** @var ResponseInterface $response */
+                        $response = $result['value'];
+                        $data = JSON::decode((string) $response->getBody(), true);
+                        $reports[] = SendReport::success($target, $data);
+                    } else {
+                        $reports[] = SendReport::failure($target, $result['reason']);
+                    }
+                }
+
+                return MulticastSendReport::withItems($reports);
+            })
+            ->wait()
+        ;
     }
 
     /**