+++ src/Firebase/Messaging/ApiClient.php
@@ -4,10 +4,13 @@ namespace Kreait\Firebase\Messaging;
 
 use GuzzleHttp\ClientInterface;
 use GuzzleHttp\Exception\RequestException;
+use GuzzleHttp\Promise\PromiseInterface;
 use GuzzleHttp\Psr7\Request;
 use Kreait\Firebase\Exception\MessagingException;
+use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\UriInterface;
+use Throwable;
 
 class ApiClient
 {
@@ -23,14 +26,28 @@ class ApiClient
 
     public function sendMessage(Message $message): ResponseInterface
     {
-        return $this->request('POST', 'messages:send', [
+        return $this->sendMessageAsync($message)->wait();
+    }
+
+    public function sendMessageAsync(Message $message): PromiseInterface
+    {
+        $request = $this->createRequest('POST', 'messages:send');
+
+        return $this->sendAsync($request, [
             'json' => ['message' => $message->jsonSerialize()],
         ]);
     }
 
     public function validateMessage(Message $message): ResponseInterface
     {
-        return $this->request('POST', 'messages:send', [
+        return $this->validateMessageAsync($message)->wait();
+    }
+
+    public function validateMessageAsync(Message $message): PromiseInterface
+    {
+        $request = $this->createRequest('POST', 'messages:send');
+
+        return $this->sendAsync($request, [
             'json' => [
                 'message' => $message->jsonSerialize(),
                 'validate_only' => true,
@@ -38,22 +55,17 @@ class ApiClient
         ]);
     }
 
-    private function request($method, $endpoint, array $options = null): ResponseInterface
+    private function sendAsync(RequestInterface $request, array $options = null): PromiseInterface
     {
         $options = $options ?? [];
 
-        $request = $this->createRequest($method, $endpoint);
-
-        try {
-            return $this->client->send($request, $options);
-        } catch (RequestException $e) {
-            throw MessagingException::fromRequestException($e);
-        } catch (\Throwable $e) {
-            throw new MessagingException($e->getMessage(), $e->getCode(), $e);
-        }
+        return $this->client->sendAsync($request, $options)
+            ->then(null, function (Throwable $e) {
+                throw $this->convertError($e);
+            });
     }
 
-    private function createRequest($method, $endpoint): Request
+    private function createRequest(string $method, string $endpoint): Request
     {
         /** @var UriInterface $uri */
         $uri = $this->client->getConfig('base_uri');
@@ -61,5 +73,14 @@ class ApiClient
         $uri = $uri->withPath($path);
 
         return new Request($method, $uri);
+    }
+
+    private function convertError(Throwable $error): MessagingException
+    {
+        if ($error instanceof RequestException) {
+            return MessagingException::fromRequestException($error);
+        }
+
+        return new MessagingException($error->getMessage(), $error->getCode(), $error);
     }
 }