+++ tests/Integration/MessagingTest.php
@@ -5,7 +5,9 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Tests\Integration;
 
 use Kreait\Firebase\Exception\Messaging\InvalidMessage;
+use Kreait\Firebase\Exception\MessagingException;
 use Kreait\Firebase\Messaging;
+use Kreait\Firebase\Messaging\CloudMessage;
 use Kreait\Firebase\Tests\IntegrationTestCase;
 
 class MessagingTest extends IntegrationTestCase
@@ -25,8 +27,9 @@ class MessagingTest extends IntegrationTestCase
         $message = MessageTestCase::createFullMessageData();
         $message['condition'] = "'dogs' in topics || 'cats' in topics";
 
-        $this->messaging->send($message);
-        $this->assertTrue($noExceptionHasBeenThrown = true);
+        $result = $this->messaging->send($message);
+
+        $this->assertArrayHasKey('name', $result);
     }
 
     public function testValidateValidMessage()
@@ -34,8 +37,9 @@ class MessagingTest extends IntegrationTestCase
         $message = MessageTestCase::createFullMessageData();
         $message['condition'] = "'dogs' in topics || 'cats' in topics";
 
-        $this->messaging->validate($message);
-        $this->assertTrue($noExceptionHasBeenThrown = true);
+        $result = $this->messaging->validate($message);
+
+        $this->assertArrayHasKey('name', $result);
     }
 
     public function testValidateInvalidMessage()
@@ -45,6 +49,31 @@ class MessagingTest extends IntegrationTestCase
 
         $this->expectException(InvalidMessage::class);
         $this->messaging->validate($message);
+    }
+
+    public function testSendMulticastWithValidAndInvalidTarget()
+    {
+        $message = CloudMessage::fromArray([]);
+        $tokens = [
+            $valid = self::$registrationTokens[0],
+            $invalid = 'invalid',
+        ];
+
+        $report = $this->messaging->sendMulticast($message, $tokens);
+
+        $this->assertCount(2, $report);
+        $this->assertTrue($report->hasFailures());
+        $this->assertCount(1, $report->failures());
+        $this->assertCount(1, $report->successes());
+
+        $success = $report->successes()->getItems()[0];
+        $this->assertSame($valid, $success->target()->value());
+        $this->assertInternalType('array', $success->result());
+        $this->assertArrayHasKey('name', $success->result());
+
+        $failure = $report->failures()->getItems()[0];
+        $this->assertSame($invalid, $failure->target()->value());
+        $this->assertInstanceOf(MessagingException::class, $failure->error());
     }
 
     public function testSubscribeToTopic()